requirements.txt
requests>=2.31.0
openai>=1.0.0
elevenlabs>=0.2.0
moviepy>=1.0.0
python-dotenv>=1.0.0
click>=8.0.0